Add a performance summary report to SLAM.
You can now request to create a SLAM performance summary.
When asked for, it will collect executions times and memory metrics.
A report string is generated by calling (PF)Slam2D::Summay::report().

A set of "Time" utilities is now available, e.g. Time, Timer, Rate.
This was created to support the creation of the performance summary
but it has a generic use when dealing with simple time actions.