[sdm] Fix a bug where some cells were not updated correctly. Save ~30% in memory usage.
I wrongly assumed that a neighbor cell already in the "lower" queue
could be safely ignored.  However, an already queued cell can be
re-queued with a higher priority due to a closer obstacle.

The coordinates to the closest obstacle is now in relative coordinates.
With this change, the information can be saved in a structure with 3
short numbers instead of 3 integers. This decreases the cell structure
from 16 bytes to 10 bytes.