[slam2d] Introduce transient mapping.
With transient mapping, only the area of the map that is covered by the
current surface (i.e. laser scan) is kept after an update.

In practice, the axis aligned bounding box (AABB) of the current surface is
tested for intersection against the AABB of all patches in the map, and
patches with an AABB that do not intersect is removed.