Fix a bug in the 'previous cache'.
On read-only a cell can return NULL, but on write it is assumed that the
cell (i.e., the patch) is properly allocated. If we read-only a cell that
is from a non-existing patch and then we immediately try to write the same
cell a segfault would occur.

This happens because of the 'previous cache'. The non-existing patch
would be cached as NULL on the read-only operation and then the write
operation would used that NULL cached patch.

This incorrect behaviour is fixed by invalidating the cache in the write
operation if the previous patch is NULL.